🐛 Fix: Update TimePicker a11y
TAB and SHIFT+TAB now work as expected when cycling time inputs
