--- packages/web/src/components/TimePicker/index.ts
@@ -1,3 +1,0 @@
-import { TimePicker } from "./TimePicker";
-
-export { TimePicker };
