+++ packages/web/src/views/Forms/EventForm/DateTimeSection/DatePickers/styled.ts
@@ -1,0 +1,6 @@
+import { Flex } from "@web/components/Flex";
+import styled from "styled-components";
+
+export const StyledDateFlex = styled(Flex)`
+  width: 120px;
+`;
