+++ packages/web/src/views/Forms/EventForm/DateTimeSection/DateTimeSection.tsx
@@ -4,17 +4,10 @@ import customParseFormat from "dayjs/plugin/customParseFormat";
 import { Categories_Event, Schema_Event } from "@core/types/event.types";
 import { SelectOption } from "@web/common/types/component.types";
 import { AlignItems } from "@web/components/Flex/styled";
-import { TimePicker } from "@web/components/TimePicker";
-import {
-  getTimeOptionByValue,
-  getTimeOptions,
-  mapToBackend,
-  shouldAdjustComplimentTime,
-} from "@web/common/utils/web.date.util";
-import { Option_Time } from "@web/common/types/util.types";
 
-import { DatePickers } from "./DatePickers";
-import { StyledDateTimeFlex, StyledTimeFlex } from "./styled";
+import { DatePickers } from "./DatePickers/DatePickers";
+import { StyledDateTimeFlex } from "./styled";
+import { TimePickers } from "./TimePicker/TimePickers";
 
 dayjs.extend(customParseFormat);
 
@@ -24,7 +17,6 @@ export interface Props {
   event: Schema_Event;
   endTime: SelectOption<string>;
   inputColor?: string;
-  nextElementRef: React.RefObject<HTMLTextAreaElement>;
   isEndDatePickerOpen: boolean;
   isStartDatePickerOpen: boolean;
   selectedEndDate: Date;
@@ -44,7 +36,6 @@ export const DateTimeSection: FC<Props> = ({
   category,
   event,
   inputColor,
-  nextElementRef,
   isEndDatePickerOpen,
   isStartDatePickerOpen,
   selectedEndDate,
@@ -59,93 +50,6 @@ export const DateTimeSection: FC<Props> = ({
   startTime,
   endTime,
 }) => {
-  const timeOptions = getTimeOptions();
-
-  const endTimeRef = useRef(null);
-
-  const adjustComplimentTimeIfNeeded = (
-    changed: "start" | "end",
-    value: string
-  ): Option_Time => {
-    const start = changed === "start" ? value : startTime.value;
-    const end = changed === "end" ? value : endTime.value;
-
-    const { shouldAdjust, adjustment, compliment } = shouldAdjustComplimentTime(
-      changed,
-      {
-        oldStart: startTime.value,
-        start,
-        oldEnd: endTime.value,
-        end,
-      }
-    );
-
-    if (shouldAdjust) {
-      if (changed === "start") {
-        const _correctedEnd = compliment.add(adjustment, "minutes");
-        const correctedEnd = getTimeOptionByValue(_correctedEnd);
-        setEndTime(correctedEnd);
-        return correctedEnd;
-      }
-
-      if (changed === "end") {
-        const _correctedStart = compliment.subtract(adjustment, "minutes");
-        const correctedStart = getTimeOptionByValue(_correctedStart);
-        setStartTime(correctedStart);
-        return correctedStart;
-      }
-    }
-
-    const defaultOption = changed === "start" ? startTime : endTime;
-    return defaultOption;
-  };
-
-  const onSelectEndTime = (option: SelectOption<string>) => {
-    setEndTime(option);
-    const correctedStart = adjustComplimentTimeIfNeeded("end", option.value);
-
-    if (endTime.value && endTime.value !== option.value) {
-      const { startDate, endDate } = mapToBackend({
-        startDate: selectedStartDate,
-        endDate: selectedEndDate,
-        startTime: correctedStart ? correctedStart : startTime,
-        endTime: option,
-        isAllDay: false,
-      });
-
-      const _event = {
-        ...event,
-        startDate,
-        endDate,
-      };
-
-      setEvent(_event);
-    }
-  };
-
-  const onSelectStartTime = (option: SelectOption<string>) => {
-    setStartTime(option);
-    const correctedEnd = adjustComplimentTimeIfNeeded("start", option.value);
-
-    if (startTime.value && startTime.value !== option.value) {
-      const { startDate, endDate } = mapToBackend({
-        startDate: selectedStartDate,
-        endDate: selectedEndDate,
-        startTime: option,
-        endTime: correctedEnd ? correctedEnd : endTime,
-        isAllDay: false,
-      });
-
-      const _event = {
-        ...event,
-        startDate,
-        endDate,
-      };
-
-      setEvent(_event);
-    }
-  };
-
   return (
     <StyledDateTimeFlex alignItems={AlignItems.CENTER} role="tablist">
       {category === Categories_Event.ALLDAY && (
@@ -164,28 +68,17 @@ export const DateTimeSection: FC<Props> = ({
       )}
 
       {category === Categories_Event.TIMED && (
-        <StyledTimeFlex alignItems={AlignItems.CENTER}>
-          <TimePicker
-            bgColor={bgColor}
-            value={startTime}
-            inputId="startTimePicker"
-            nextElementRef={endTimeRef}
-            onChange={onSelectStartTime}
-            openMenuOnFocus
-            options={timeOptions}
-          />
-          -
-          <TimePicker
-            bgColor={bgColor}
-            value={endTime}
-            inputId="endTimePicker"
-            nextElementRef={nextElementRef}
-            onChange={onSelectEndTime}
-            openMenuOnFocus
-            options={timeOptions}
-            ref={endTimeRef}
-          />
-        </StyledTimeFlex>
+        <TimePickers
+          bgColor={bgColor}
+          endTime={endTime}
+          event={event}
+          setStartTime={setStartTime}
+          setEndTime={setEndTime}
+          setEvent={setEvent}
+          startTime={startTime}
+          selectedEndDate={selectedEndDate}
+          selectedStartDate={selectedStartDate}
+        />
       )}
     </StyledDateTimeFlex>
   );
