+++ packages/web/src/views/Forms/EventForm/DateTimeSection/TimePicker/TimePicker.tsx
@@ -1,62 +1,51 @@
-import React, { ForwardedRef, forwardRef, useRef, useState } from "react";
-import ReactSelect, { Props as RSProps, SelectInstance } from "react-select";
+import React from "react";
+import ReactSelect, { Props as RSProps } from "react-select";
 import { Key } from "ts-key-enum";
 import { SelectOption } from "@web/common/types/component.types";
 import { Option_Time } from "@web/common/types/util.types";
 
 import { StyledTimePicker, StyledDivider } from "./styled";
 
-export interface Props extends Omit<RSProps, "value"> {
+export interface Props extends Omit<RSProps, "onChange" | "value"> {
   bgColor: string;
-  nextElementRef?: React.RefObject<HTMLTextAreaElement>;
+  isMenuOpen: boolean;
   onChange: (option: SelectOption<string>) => void;
   options?: Option_Time[];
   selectClassName?: string;
+  setIsMenuOpen: React.Dispatch<React.SetStateAction<boolean>>;
   value: SelectOption<string>;
 }
 
-export const _TimePicker = (
-  {
-    bgColor,
-    nextElementRef,
-    onChange: _onChange,
-    options,
-    selectClassName,
-    value,
-    ...props
-  }: Props,
-  ref: ForwardedRef<ReactSelect>
-  // ref: Ref<SelectInstance>
-) => {
-  const [isFocused, setIsFocused] = useState(false);
-  const [isMenuOpened, setIsMenuOpen] = useState(false);
-
+export const TimePicker = ({
+  bgColor,
+  isMenuOpen,
+  onChange: _onChange,
+  options,
+  selectClassName,
+  setIsMenuOpen,
+  value,
+  ...props
+}: Props) => {
   const TIMEPICKER = "timepicker";
   let scrollTimer: number;
 
   return (
-    <StyledTimePicker bgColor={bgColor} isOpen={isMenuOpened}>
+    <StyledTimePicker bgColor={bgColor}>
       <ReactSelect
         {...props}
-        ref={ref}
         className={selectClassName}
         classNamePrefix={TIMEPICKER}
         value={value}
         maxMenuHeight={4 * 41}
-        noOptionsMessage={() => "Nothin'. Typo?"}
-        onBlur={() => {
-          console.log("blurred");
-          setIsFocused(false);
-        }}
-        onFocus={() => {
-          console.log("focused");
-          setIsFocused(true);
-        }}
+        blurInputOnSelect
+        menuIsOpen={isMenuOpen}
+        //@ts-expect-error uses custom onChange to manage focus in parent
         onChange={_onChange}
         onKeyDown={(e) => {
           const key = e.key;
 
           if (key === Key.Enter || key === Key.Backspace) {
+            console.log("stopping prop");
             e.stopPropagation();
           }
 
@@ -68,6 +57,10 @@ export const _TimePicker = (
             setIsMenuOpen(false);
             e.stopPropagation();
           }
+
+          if (key === Key.Tab) {
+            setIsMenuOpen(false);
+          }
         }}
         onMenuOpen={() => {
           scrollTimer = window.setTimeout(() => {
@@ -78,29 +71,16 @@ export const _TimePicker = (
               defaultOpt.scrollIntoView();
             }
           }, 15);
-
           setIsMenuOpen(true);
         }}
         onMenuClose={() => {
-          console.log("closed menu");
-          setIsFocused(false);
-          setIsMenuOpen(false);
           clearTimeout(scrollTimer);
-
-          if (nextElementRef && nextElementRef.current) {
-            console.log("focusing on next input...");
-            nextElementRef.current.focus();
-          } else {
-            console.log("doing nothing with refs");
-          }
         }}
         openMenuOnFocus={true}
         options={options}
         tabSelectsValue={false}
       />
-      {isFocused && <StyledDivider width="calc(100% - 16px)" />}
+      {isMenuOpen && <StyledDivider width="calc(100% - 16px)" />}
     </StyledTimePicker>
   );
 };
-
-export const TimePicker = forwardRef(_TimePicker);
